MockVehicle: Ignore location callback in first takeoff update
Do not call the callback for the location update until after the takeoff
actually made a nonzero update. This is to prevent the location check
from assuming the vehicle flew through an object when it takes off from
the plane, which may be an object from the VRML scene. Ignore this so
that we can use the location update for simulated drone flights; it
previously only worked when the operating altitude is 0, i.e., the
vehicle is a rover.