Simulations: Determine altitudes per scene
Different scenes require different altitudes to operate on: the castle
has usually been run at 4 meters, but the trees in trees_river are not
detectable at that altitude and the deranged_house is lower than that,
causing the vehicle to fly through an object. Thus, determine the
altitude in a similar setting as the translation.