Simulations: Also test space size
This parameter also influences the results. Actually the space size
depends on the scene, but certain scenes (castle) may work with
different sizes, so rerun them anyway. Keep the old runs if they exist
by fixing the old path name and checking if that exists, moving it to
the new name.

Also clean up some variable names and debug print.