Vehicle: Introduce target location property
Make it possible to request what location the vehicle is going toward.
This target location should be in a straight line from the "starting"
location, but vehicles may diverge. Some vehicles and modes are unable
to provide this and may instead return the current location or the next
waypoint (even if it is not in a straight line).

This is useful information for the mission monitoring components, since
we may need to adjust the waypoints to follow a safe path in case of
oncoming collisions. For GUIDED mode, the mission itself is mostly
responsible for keeping the vehicle safe, but for AUTO mode this was
always cumbersome since the autopilot is busy following the mission.
This is a step toward using `AStar` in between the mission if necessary.