Collision avoidance: Configurable unsafe path cost
Make it possible to change the distance for routes for which there
exists no guaranteed safe path. By default this is infinity, thus
causing the individual to become infeasible. Because the planning
algorithm parameters can cause many individuals to have conflicting
waypoints, and we may be able to solve the conflict during the mission
on the vehicles, the cost should be configurable.

Take the original norm distance into account, and add the unsafe path
cost setting to it. Clean up some comments as well.

Aside from this, revamp the settings widgets to support inputting
a value for infinity. This is only possible through a context menu for
float type setting that explicitly allow it and do not have a min/max
constraint. The Qt validator for floating point numbers disallows
inf/-inf input, even though it could theoretically support them. Thus we
refactor the use of the validator state for checking if the value is
allowed, including considering the default always allowable.

Additionally, the settings table widget now formats the help text such
that it spans across multiple lines, one for each sentence. Since the
unsafe path cost setting text is quite long, this makes the tooltip more
readable on non-ultra-wide screens.