Control panel settings widgets: Split large file
The settings widgets were becoming quite large and were getting close to
the limit imposed by pylint upon the number of lines in a module. To get
away from this, split the settings widgets and the form widgets into two
separate files. We could split each class into its own module, but there
is no use for that at the moment.