Control panel settings Float form widget: Always allow infinite
The infinite value, most notably used for the unsafe path cost, can be
inserted via the context menu, but when the value is already default,
problems may arise. Due to this, it may appear as if the value was
accepted, but in fact the settings are ignored. This is because the
planning view checks for disallowed values while updating the values,
and this means that the validator state may be put in a disallowed state
even though the value was accepted.

This ensures that the infinite values are always allowed if they should
be, so that default values do not lead to quasi-rejected states.