Settings: Handle defaults for tuple types better
The settings defaults are imported from JSON, which only knows arrays
and has no difference for lists and tuples. Thus the settings with tuple
type would actually have a default value of list type. This is
problematic when we want to check if the current value is not default,
because the `Arguments` parser would cast the value to a tuple.

Note that the control panel also used lists for tuples, given its focus
on writing JSON settings overrides. Create a new tuple form widget that
does the same as the list form widget, except cast the value.

This change is limited to the main values, not the tuple type subvalues
within lists or dictionaries. This is because it would be too cumbersome
to detect and convert these partial values as well, and `Arguments` does
not provide much support for them either (since it does not know where
elements of a list of tuples end and does not understand dictionaries).
This does mean we have to exclude those from the tuple form widget's
casting in the case that it is used as a subwidget.