Planning experiment runner: Initial work
This is an initial implementation of the experiment runner, which is
able to execute several experiments of the same type at once. It can
pass through settings, environment, and python flags, which may be
useful in some contexts. The experiments themselves are not yet fully
fleshed out, although it does run a bunch of them and provides results
in a directory structure. The results are also not yet used anywhere.