Planning experiments: Combine multiple arguments
Make it possible to run experiments with multiple arguments. Use all
combinations in such experiments.

Also add the settings that are overridden from the command line to the
path name so that we can trace back what the exact arguments were, and
avoid conflicts when combining multiple runs in one results directory.

Ensure all exceptions are displayed in the output.