Collision avoidance: Clip home location setting
Ensure that the home location is within the bounds of the network
(including padding), since we would otherwise go out of bounds in
`Memory_Map` and `AStar`, thus failing the planning. We should be
cautious with this setting because it cannot be changed using the
command-line arguments, which means that certain experiments with
small-size networks fail. This is just a stop gap, since the home
location should just be fixed in the first place, and it might be
replaced with an optimizable variable at some point.