Planning experiments: Initial results gathering
We collect results by extracting information from the output logs and
making this available in a usable format. We may later extract more
information and extract from other result data, but this already
provides a basis for making summarizable results.

The results collection class shares a common ground with the experiment
runner, so a new `Experiment` class is created that wraps a lot of the
reusable methods and data. The list of settings components that is
available in the control panel happens to be the same as the components
used in the experiments, so move them to a common variable too.

The results gathering works for both the experiments in the JSON as well
as singular experiments that are deducible from that file. It can
extract from different paths.