Let each vehicle track the other vehicle locations
This makes the `RF_Sensor` pass along the location from the
"rssi_broadcast" packet to the valid callback, such that the Environment
can store this information. The `Monitor` can then add the locations to
the memory map, so that it can know where the different vehicles are.
This could be used later on to optionally halt the vehicle and find
a new safe path in case the current route crosses any other vehicle.

Simplify the required sensors in `Environment` such that it always
contains the own sensor ID. Finally, document the `Monitor` some more.